[MINOR] Fix static code analysis warnings, part II
This patch applies additional changes from an automated static code
analysis. These changes include a number of minor bugs in tests and
system code. The most important bug fix, is the inactive IPA early
abort feature (avoiding unnecessary compilation overhead).

+++ src/main/java/org/apache/sysds/runtime/compress/lib/CLALibCompAgg.java
@@ -345,7 +345,7 @@ public final class CLALibCompAgg {
 	}
 
 	private static void divideByNumberOfCellsForMeanCols(CompressedMatrixBlock m1, MatrixBlock ret) {
-		double div = (double) m1.getNumRows();
+		double div = m1.getNumRows();
 
 		if(ret.isInSparseFormat()) {
 			SparseBlock sb = ret.getSparseBlock();
